refactor(utils,scripts): low-priority polish and consistency cleanups
Breaking (script path): scripts/xy/compare_correlations.py is renamed to
scripts/xy/correlation_comparison.py, matching its Ising and Clock
siblings; outputs unchanged. References in SCRIPTS.md, the BKT notebook,
and the generated API docs are updated.

utils:
- utils/__init__ exports the actual public surface instead of a stale
  eleven-name subset.
- calculate_autocorr computes a real power spectrum and normalizes only
  the consumed lags; summarize_entropy_observable builds its result dict
  through one helper; the chi/Cv fluctuation formulas live once in
  utils.observables.derived_thermo_estimate, shared with the derived
  summarizer; leftover blank-line blocks removed.
- Six public statistics functions gain full Parameters/Returns sections.
- setup_logging can attach a log file on a later call (console and file
  handlers guarded independently, duplicate file handlers skipped);
  parallel_sweep signature uses parameterized generics.

scripts:
- --seed on the three ordering_evolution scripts (default stays
  unseeded); --n-seeds accepted as canonical spelling on the kinetics
  scripts (--seeds keeps working).
- coarsening_analysis: fixed a NameError on empty --quench-fracs
  (loop-variable leakage), offset the bridge seed block by 300 so
  crossover trajectories are independent of the frac=0.5 quench traces,
  exposed the four hardcoded xi_eq parameters as --xi-eq-* flags, and
  delegated the 1/e correlation length to the shared interpolating
  helper.
- --max-seed-attempts help corrected to per-temperature; the
  max(n, 10*n) no-op simplified; clock correlation --seed help matches
  behavior; bkt_transition accumulates vortex densities in an array.

Validation: uv run pytest (285 passed), ruff, mypy, pre-push hooks green.

+++ utils/kinetics_helpers.py
@@ -44,11 +44,11 @@ def compute_mean_intercept_length(*, sim: Any) -> float:
         return 0.0
     spins = sim.spins.astype(np.int8)
     N = sim.size
-    row_walls = np.sum(spins[:, :-1] * spins[:, 1:] < 0)
-    col_walls = np.sum(spins[:-1, :] * spins[1:, :] < 0)
+    row_walls = int(np.sum(spins[:, :-1] * spins[:, 1:] < 0))
+    col_walls = int(np.sum(spins[:-1, :] * spins[1:, :] < 0))
     row_wrap = int(np.sum(spins[:, -1] * spins[:, 0] < 0))
     col_wrap = int(np.sum(spins[-1, :] * spins[0, :] < 0))
-    total_walls = int(row_walls) + int(col_walls) + row_wrap + col_wrap
+    total_walls = row_walls + col_walls + row_wrap + col_wrap
     mean_walls = total_walls / (2 * N)
     return float(N) / mean_walls if mean_walls != 0.0 else float(N)
